// libs/in-app/src/lib/models/accident.ts
export class Accident {
  AccidentID: number;
  AccidentPrepend: string;
  AccidentTypeID: number = 27;
  AccidentTypeName: string;
  ArabicAccidentTypeName: string;
  ResponsibilityTypeID: number = 41;
  ResponsibilityTypeName: string;
  ArabicResponsibilityTypeName: string;
  AccidentNo: string;
  tempAccidentNo: string;
  VIN: string;
  CarsInvolved: number;
  AccidentHappendOn: any |Date;
  PlateNo: string;
  MakeID: number;
  ModelID: number;
  ProductionYear: number;
  CompanyID: number;
  ICName: string;
  CPPhone: string;
  CPName: string;
  CreatedOn: Date;
  CreatedBy: number;
  ModifiedBy: number;
  ModifiedOn: Date;
  ImgURL: string;
  IsDeleted: boolean;
  AccidentCreatedOn: Date;
  YearCode: string;
  MakeName: string;
  ModelCode: string;
  VehicleOwnerName: string;
  OwnerPhoneNo: string;
  ICWorkshopID: number;
  WorkshopName: string;
  CarLicenseFront: string;
  CarLicenseBack: string;
  OwnerIDFront: string;
  OwnerIDBack: string;
  PoliceReport: string;
  PDFReport: string;
  StatusID: number = 20;
  UserName: string;
  InprogressRequestCount: number

  CarLicenseFrontDescription: string;
  CarLicenseBackDescription: string;
  OwnerIDFrontDescription: string;
  OwnerIDBackDescription: string;
  PoliceReportDescription: string;
  PDFReportDescription: string;

  EngineTypeID: number;
  BodyTypeID: number = 51;
  ImportantNote: string;
  EngineTypeName: string;
  EngineTypeArabicName: string;
  BodyTypeName: string;
  BodyTypeArabicName: string;
  BodyTypeIcon: string;
  FaultyCompanyName: string;
  AccidentCreatedBy: Date;
  TotalRequestCount: number;
  DemandCount: number;
  ArabicBodyTypeName: string;



  VernishPrice: number = 0;
  MechanicLabourPrice: number = 0;
  DamagePrice: number = 0;
  TotalLabourCost: number = 0;
  TotalLabourTime: number = 0;
  FinalPartsAmount: number = 0;
  FinalLabourCost: number = 0;
  TotalPartsAmount: number = 0;
  GasPrice: number = 0;
  TotalAccidentCost: number = 0;
  FinalAccidentCost: number = 0;
  FinalLabourTime: number = 1;
  TotalOldPartsAmount: number = 0;
  RentPrice: number = 8;
  TotalMechanicLabourPrice: number = 0;

  ClearanceModifiedByName: string;
  ClearanceModifiedOn: Date;
  ClearanceCreatedOn: Date;
  ClearanceModifiedBy: number;
  ClearanceCreatedByName: string;
  ClearanceSummaryPdfUrl: string;
  IsPurchasing: boolean = false;
  ClearanceRoute: string;
  Saving: number;
  EventDateTime: Date;
  ModifiedByName: string;
  CreatedByName: string;
  CreatedByEmail: string;
  ModifiedByEmail: string;
  ArabicMakeName: string;
  ArabicModelName: string;
  WorkshopDetails: string;
  IsClearance: number;
  PendingRequestID: number;
  CloseReason: string;
  LogoURL: string;
  JCSeriesID: string;
  GroupName: string;
  EnglishMakeName: string;
  EnglishModelName: string;
  ImageCaseEncryptedName: string;
  ImageCaseURL: string;
  JCSeriesCode: string;
  AskForPartCondition: boolean;
  IsIndividualReturn: boolean = false;
  IndividualReturnID: number;
  IndividualReturnText: string;
  ShowText: boolean = false;
  TechnicalNotesStatus: number;
  UserID: number;
  faultyPlateNo: string;
  TotalAmountInWritten: string;
  FaultyPartyName: string;
  AccidentLocation: string;
  AccidentConsequences: string;
  InsuranceDocumentNo: string;
  RequestStatusName: string;
  RequestArabicStatusName: string;
  PricingType:string;
  POAmount:number;
  ChequeFreeText:string;
  ChequeDate:Date;
  ChequeCreatedOn:Date;
  ChecqueApprovalSignature:string;
  ChequePdfUrl:string;
  SignaturedBy:number;
  SignaturedByName:string;
  OtherPlusText: string;
  OtherMinText: string;
  BrokerName: string;
  ClaimentID: number;
  OurResponsibility: number;
  IsAccidentExist:boolean =false;
  DraftID:number;
  RequestCount:number;
  RequestPartCount:number;
  PublishedRequestCount:number;
  UnPublishedRequestCount:number;
  QuotationCount:number;
  MatchingOffer:number;
  NotMatchingOffers:number;
  IsAgencyAccident: boolean;
}