// libs/in-app/src/lib/models/model.ts
export class Model
{
    LanguageModelID: number;
    ModelCode: string;
    ArabicModelName:string;
    MakeID: number;
    YearCode: number;
    LanguageID: number;
    ModelID: number;
    ModelSearch: string;
    GroupName: string;
    IsSelected: boolean =false;
    ModelGroupName:string;
    MakeName: string;
    ArabicMakeName: string;
    ImgURL: string;
    UserID: number;
    AccidentCount: number;
}